refactor: extract buildRuntimeConfig to resolve.ts (DRY)
Move the 20+ field RuntimeConfig construction from loadConfig into a
dedicated buildRuntimeConfig function in resolve.ts. loadConfig now
focuses on loading/validation; resolve.ts owns all resolution logic.
